move linter package to azurerm/internal/tools
