+++ azurerm/internal/tools/azurermproviderlint/passes/AZRMR001/testdata/src/a/main.go.golden
