+++ azurerm/internal/tools/azurermproviderlint/passes/AZRMS001/testdata/src/a/main.go.golden
