+++ azurerm/internal/tools/azurermproviderlint/passes/AZRMS001/testdata/src/a/vendor
@@ -1,0 +1,1 @@
+../../../../../../../../../vendor
