--- azurermproviderlint/passes/AZRMS001/testdata/src/a/vendor
@@ -1,1 +1,0 @@
-../../../../../../vendor
